Arduino serial send sketch
Here is the sketch to upload to the Arduino to tell you if the pressure plates are activated.
Pins 8 and 9 register activation if they're shorted to ground, so the pressure plates can trigger a short to ground and do not need external pullup resistors.
The Arduino sends a single byte over and over on serial at 115200 baud that encodes the status of the pressure plates on the first 2 bits.